* g++-warning-fixes:
  BSE: use std::atomic<>, since Rapicorn's Atomic<> is gone
  LAUNCHERS: abort if seteuid/setreuid failed to drop priviledges
  TOOLS: fix signedness compiler warnings
  SHELL: silence compiler
  BEAST: fix signedness compiler warnings
  BEAST: silence compiler
  GXK: fix warning about macro redefinition
  GXK: fix compiler warnings about switch type
  GXK: fix signedness compiler warnings
  PLUGINS: fix signedness compiler warnings
  BSE: silence compiler
  BSE: fix procedure prototype warnings
  BSE: fix signedness compiler warnings
  BSE: fix compiler warnings about switch type
  BSE: fix prototype ordering in BSE_DUMMY_TYPE
  BSE: fix use of deprecated g_source_get_current_time()
  BSE: fix bse_cxx_init namespace

+++ beast-gtk/bstasciipixbuf.cc
@@ -60,7 +60,7 @@ bst_ascii_pixbuf_new (gchar character,
     char_width = width;
   if (!char_height)
     char_height = gdk_pixbuf_get_height (pixbuf);
-  if (char_width != width || char_height != gdk_pixbuf_get_height (pixbuf))
+  if (char_width != width || char_height != (uint) gdk_pixbuf_get_height (pixbuf))
     {
       GdkPixbuf *tmp_pixbuf = pixbuf;
